LJpegDecompressor/LJpegPlain code reorganization
The former LJpegPlain code was basically dealing with decoding LJpeg
blobs from Cr2 files and from Dng files. There are a number of special
cases and features that needed to be dealt with to support both use
cases. It turned out that this set of features/quirks was almost
completely disjoint:

Cr2:
* cr2_slice handling
* 80D frame overflow
* canon double width / half height
* 6D width/height swap
* mRaw/sRaw (YCrCb super sampling variants)

Dng:
* tiled decoding (offX/offY)
* decoding of data that does not fit inside the image frame

This led me split the code in two separate decompressor classes that are
a lot easier to understand than the mix of all those special case
handlings before. One is called Cr2Decompressor, the other just
LJpegDecompressor. The former LJpegDecompressor has been renamed to
AbstractLJpegDecompressor.

I also added the required parameters that were formerly set via public
members as parameters to the decode method. This interface removes
opportunities to use the classes in an unsupported way.

The diffstat numbers are almost entirely caused by the file renaming.
The decodeN_... function has only been split in two.
