+++ src/librawspeed/decompressors/LJpegDecompressor.cpp
@@ -1,7 +1,6 @@
 /*
     RawSpeed - RAW file decoder.
 
-    Copyright (C) 2009-2014 Klaus Post
     Copyright (C) 2017 Axel Waggershauser
 
     This library is free software; you can redistribute it and/or
@@ -20,19 +19,15 @@
 */
 
 #include "decompressors/LJpegDecompressor.h"
-#include "common/Common.h"                // for uint32, getHostEndianness
-#include "common/Point.h"                 // for iPoint2D
-#include "decoders/RawDecoderException.h" // for ThrowRDE
-#include "decompressors/HuffmanTable.h"   // for HuffmanTable
-#include "io/ByteStream.h"                // for ByteStream
-#include "io/IOException.h"               // for IOException
-#include <array>                          // for array
-#include <memory>                         // for unique_ptr, allocator
-#include <vector>                         // for vector
+#include "common/Common.h"
+#include "io/BitPumpJPEG.h"
+#include "io/ByteStream.h"
+
+using namespace std;
 
 namespace RawSpeed {
 
-void LJpegDecompressor::decode(uint32 offsetX, uint32 offsetY) {
+void LJpegDecompressor::decode(uint32 offsetX, uint32 offsetY, bool fixDng16Bug_) {
   if ((int)offsetX >= mRaw->dim.x)
     ThrowRDE("LJpegDecompressor: X offset outside of image");
   if ((int)offsetY >= mRaw->dim.y)
@@ -40,151 +35,73 @@ void LJpegDecompressor::decode(uint32 offsetX, uint32 offsetY) {
   offX = offsetX;
   offY = offsetY;
 
-  if (getNextMarker(false) != M_SOI)
-    ThrowRDE("LJpegDecompressor: Image did not start with SOI. Probably not an LJPEG");
+  fixDng16Bug = fixDng16Bug_;
 
-  JpegMarker m;
-  do {
-    m = getNextMarker(true);
+  AbstractLJpegDecompressor::decode();
+}
 
-    switch (m) {
-    case M_DHT:  parseDHT(); break;
-    case M_SOF3: parseSOF(&frame); break;
-    case M_SOS:  parseSOS(); break;
-    case M_DQT:  ThrowRDE("LJpegDecompressor: Not a valid RAW file.");
-    default:  // Just let it skip to next marker
+void LJpegDecompressor::decodeScan()
+{
+  if (predictorMode != 1)
+    ThrowRDE("LJpegDecompressor: Unsupported predictor mode");
+
+  for (uint32 i = 0; i < frame.cps;  i++)
+    if (frame.compInfo[i].superH != 1 || frame.compInfo[i].superV != 1)
+      ThrowRDE("LJpegDecompressor: Unsupported subsampling");
+
+  if (frame.cps == 2)
+    decodeN<2>();
+  else if (frame.cps == 3)
+    decodeN<3>();
+  else if (frame.cps == 4)
+    decodeN<4>();
+  else
+    ThrowRDE("LJpegDecompressor: Unsupported number of components");
+}
+
+// N_COMP == number of components (2, 3 or 4)
+
+template <int N_COMP>
+void LJpegDecompressor::decodeN()
+{
+  auto ht = getHuffmanTables<N_COMP>();
+  auto pred = getInitialPredictors<N_COMP>();
+  auto predNext = pred.data();
+
+  BitPumpJPEG bitStream(input);
+
+  for (unsigned y = 0; y < frame.h; ++y) {
+    auto destY = offY + y;
+    // A recoded DNG might be split up into tiles of self contained LJpeg
+    // blobs. The tiles at the bottom and the right may extend beyond the
+    // dimension of the raw image buffer. The excessive content has to be
+    // ignored. For y, we can simply stop decoding when we reached the border.
+    if (destY >= (unsigned)mRaw->dim.y)
       break;
+
+    auto dest = (ushort16*)mRaw->getDataUncropped(offX, destY);
+
+    copy_n(predNext, N_COMP, pred.data());
+    // the predictor for the next line is the start of this line
+    predNext = dest;
+
+    unsigned width = min(frame.w,
+                         (mRaw->dim.x - offX) / (N_COMP / mRaw->getCpp()));
+
+    // For x, we first process all pixels within the image buffer ...
+    for (unsigned x = 0; x < width; ++x) {
+      unroll_loop<N_COMP>([&](int i) {
+        *dest++ = pred[i] += ht[i]->decodeNext(bitStream);
+      });
     }
-  } while (m != M_EOI);
-}
-
-void LJpegDecompressor::parseSOF(SOFInfo* sof) {
-  uint32 headerLength = input.getShort();
-  sof->prec = input.getByte();
-  sof->h = input.getShort();
-  sof->w = input.getShort();
-  sof->cps = input.getByte();
-
-  if (sof->prec > 16)
-    ThrowRDE("LJpegDecompressor: More than 16 bits per channel is not supported.");
-
-  if (sof->cps > 4 || sof->cps < 1)
-    ThrowRDE("LJpegDecompressor: Only from 1 to 4 components are supported.");
-
-  if (headerLength != 8 + sof->cps*3)
-    ThrowRDE("LJpegDecompressor: Header size mismatch.");
-
-  for (uint32 i = 0; i < sof->cps; i++) {
-    sof->compInfo[i].componentId = input.getByte();
-    uint32 subs = input.getByte();
-    frame.compInfo[i].superV = subs & 0xf;
-    frame.compInfo[i].superH = subs >> 4;
-    uint32 Tq = input.getByte();
-    if (Tq != 0)
-      ThrowRDE("LJpegDecompressor: Quantized components not supported.");
-  }
-  sof->initialized = true;
-
-  mRaw->metadata.subsampling.x = sof->compInfo[0].superH;
-  mRaw->metadata.subsampling.y = sof->compInfo[0].superV;
-}
-
-void LJpegDecompressor::parseSOS() {
-  if (!frame.initialized)
-    ThrowRDE("LJpegDecompressor::parseSOS: Frame not yet initialized (SOF Marker not parsed)");
-
-  uint32 headerLength = input.getShort();
-  if (headerLength != 3 + frame.cps * 2 + 3)
-    ThrowRDE("LJpegDecompressor::parseSOS: Invalid SOS header length.");
-
-  uint32 soscps = input.getByte();
-  if (frame.cps != soscps)
-    ThrowRDE("LJpegDecompressor::parseSOS: Component number mismatch.");
-
-  for (uint32 i = 0; i < frame.cps; i++) {
-    uint32 cs = input.getByte();
-    uint32 td = input.getByte() >> 4;
-
-    if (td > 3 || !huff[td])
-      ThrowRDE("LJpegDecompressor::parseSOS: Invalid Huffman table selection.");
-
-    int ciIndex = -1;
-    for (uint32 j = 0; j < frame.cps; ++j) {
-      if (frame.compInfo[j].componentId == cs)
-        ciIndex = j;
+    // ... and discard the rest.
+    for (unsigned x = width; x < frame.w; ++x) {
+      unroll_loop<N_COMP>([&](int i) {
+        ht[i]->decodeNext(bitStream);
+      });
     }
-
-    if (ciIndex == -1)
-        ThrowRDE("LJpegDecompressor::parseSOS: Invalid Component Selector");
-
-    frame.compInfo[ciIndex].dcTblNo = td;
   }
-
-  // Get predictor
-  predictorMode = input.getByte();
-  if (predictorMode > 8)
-    ThrowRDE("LJpegDecompressor::parseSOS: Invalid predictor mode.");
-
-  input.skipBytes(1);         // Se + Ah Not used in LJPEG
-  Pt = input.getByte() & 0xf; // Point Transform
-
-  decodeScan();
-}
-
-void LJpegDecompressor::parseDHT() {
-  uint32 headerLength = input.getShort() - 2; // Subtract myself
-
-  while (headerLength)  {
-    uint32 b = input.getByte();
-
-    uint32 htClass = b >> 4;
-    if (htClass != 0)
-      ThrowRDE("LJpegDecompressor::parseDHT: Unsupported Table class.");
-
-    uint32 htIndex = b & 0xf;
-    if (htIndex >= huff.size())
-      ThrowRDE("LJpegDecompressor::parseDHT: Invalid huffman table destination id.");
-
-    if (huff[htIndex] != nullptr)
-      ThrowRDE("LJpegDecompressor::parseDHT: Duplicate table definition");
-
-    auto ht = make_unique<HuffmanTable>();
-
-    // copy 16 bytes from input stream to number of codes per length table
-    uint32 nCodes = ht->setNCodesPerLength(input.getBuffer(16));
-    // spec says 16 different codes is max but Hasselblad violates that -> 17
-    if (nCodes > 17 || headerLength < 1 + 16 + nCodes)
-      ThrowRDE("LJpegDecompressor::parseDHT: Invalid DHT table.");
-
-    // copy nCodes bytes from input stream to code values table
-    ht->setCodeValues(input.getBuffer(nCodes));
-
-    // see if we already have a HuffmanTable with the same codes
-    for (const auto& i : huffmanTableStore)
-      if (*i == *ht)
-        huff[htIndex] = i.get();
-
-    if (!huff[htIndex]) {
-      // setup new ht and put it into the store
-      ht->setup(mFullDecodeHT, mDNGCompatible);
-      huff[htIndex] = ht.get();
-      huffmanTableStore.emplace_back(std::move(ht));
-    }
-    headerLength -= 1 + 16 + nCodes;
-  }
-}
-
-JpegMarker LJpegDecompressor::getNextMarker(bool allowskip) {
-  uchar8 c0, c1 = input.getByte();
-  do {
-    c0 = c1;
-    c1 = input.getByte();
-  } while (allowskip && !(c0 == 0xFF && c1 != 0 && c1 != 0xFF));
-
-  if (!(c0 == 0xFF && c1 != 0 && c1 != 0xFF))
-    ThrowRDE("LJpegDecompressor::getNextMarker: (Noskip) Expected marker not found. Propably corrupt file.");
-
-  return (JpegMarker)c1;
+  input.skipBytes(bitStream.getBufferPosition());
 }
 
 } // namespace RawSpeed
